requirements_final.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
dnspython==2.6.1
docker==7.1.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
flake8==7.0.0
Flask==3.0.3
fonttools==4.52.4
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
loguru==0.7.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mlflow==2.13.1
mypy-extensions==1.0.0
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
pyarrow==15.0.2
pycodestyle==2.11.1
pydantic==2.7.2
pydantic_core==2.18.3
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyyaml_env_tag==0.1
querystring-parser==1.2.4
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
-e git+https://github.com/Darshan31Patel/NYC-TAXI.git@b6e75cee8c557fa79947083feaccda6ba0ca59f5#egg=src
starlette==0.37.2
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
waitress==3.0.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wrapt==1.16.0
zipp==3.19.0
